Remove APIPublisher
