+++ Azure.ApiManagement.Client/Azure.ApiManagement.Client.csproj
@@ -167,8 +167,6 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ClientProxy\APIPublisher.cs" />
-    <Compile Include="ClientProxy\APIConfiguration.cs" />
     <Compile Include="DataModel\Abstract\EntityIdGenerator.cs" />
     <Compile Include="DataModel\Abstract\UpdateEntityBase.cs" />
     <Compile Include="DataModel\Exceptions\SwaggerResourceException.cs" />
@@ -203,7 +201,7 @@
     <Compile Include="Swagger\Models\Header.cs" />
     <Compile Include="Swagger\Models\Info.cs" />
     <Compile Include="Swagger\AbstractSwaggerReader.cs" />
-    <Compile Include="ClientProxy\APIComposer.cs" />
+    <Compile Include="ClientProxy\APIBuilder.cs" />
     <Compile Include="Swagger\Models\Item.cs" />
     <Compile Include="Swagger\Models\License.cs" />
     <Compile Include="Swagger\Models\OperationMethod.cs" />
@@ -219,7 +217,7 @@
     <Compile Include="Swagger\Models\Tag.cs" />
     <Compile Include="Swagger\SwaggerUrlReader.cs" />
     <Compile Include="Swagger\Models\Xml.cs" />
-    <Compile Include="ClientProxy\APIBuilder.cs" />
+    <Compile Include="ClientProxy\APISwaggerBuilder.cs" />
     <Compile Include="Utilities\Constants.cs" />
     <Compile Include="Utilities\PrintMessage.cs" />
     <Compile Include="Utilities\Utility.cs" />
